requirements.txt
asgiref==3.8.1
awscrt==0.22.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycryptodomex==3.21.0
PyJWT==2.8.0
#pylint==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.0
tomlkit==0.13.2
typing_extensions==4.12.0
tzdata==2024.1
urllib3>=1.26,<3.0
